% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_trial_module.R
\name{analyze_clinical_trial_data}
\alias{analyze_clinical_trial_data}
\title{Analyze Clinical Trial Data}
\usage{
analyze_clinical_trial_data(clinical_trial_data)
}
\arguments{
\item{clinical_trial_data}{A dataframe containing preprocessed clinical trial data.}
}
\description{
This function performs basic analysis on clinical trial data, such as summary statistics and distribution plots.
}
