% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_models_module.R
\name{build_predictive_model}
\alias{build_predictive_model}
\title{Build Predictive Model}
\usage{
build_predictive_model(training_data)
}
\arguments{
\item{training_data}{A dataframe containing training data.}
}
\value{
A trained predictive model.
}
\description{
This function builds a predictive model for patient outcome predictions.
}
