% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_models_module.R
\name{evaluate_predictive_model}
\alias{evaluate_predictive_model}
\title{Evaluate Predictive Model}
\usage{
evaluate_predictive_model(trained_model, test_data)
}
\arguments{
\item{trained_model}{A trained predictive model.}

\item{test_data}{A dataframe containing test data.}
}
\value{
A list of evaluation metrics.
}
\description{
This function evaluates the performance of a predictive model.
}
\examples{
# Example usage:
# Assuming you have a trained model 'trained_model' and a test data frame 'test_data'
evaluation_result <- evaluate_predictive_model(trained_model, test_data)
}
