% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical_imaging_module.R
\name{interpret_radiological_image}
\alias{interpret_radiological_image}
\title{Interpret Radiological Images}
\usage{
interpret_radiological_image(radiological_image)
}
\arguments{
\item{radiological_image}{A radiological image for interpretation.}
}
\description{
This function provides guidance on interpreting radiological images.
}
