% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_trial_module.R
\name{load_clinical_trial_data}
\alias{load_clinical_trial_data}
\title{Clinical Trial Module}
\usage{
load_clinical_trial_data(path)
}
\arguments{
\item{path}{The file path to the CSV file containing clinical trial data.}
}
\value{
A dataframe containing clinical trial data.
}
\description{
This module focuses on understanding and analyzing clinical trial data,
including trial outcomes, patient profiles, and treatment efficacy.
}
\examples{
clinical_data <- load_clinical_trial_data("path/to/clinical_data.csv")
}
\author{
Khawla Aghzawi
}
\keyword{CSV}
\keyword{Clinical}
\keyword{Data}
\keyword{Load}
\keyword{This}
\keyword{Trial}
\keyword{a}
\keyword{analysis}
\keyword{and}
\keyword{as}
\keyword{clinical}
\keyword{containing}
\keyword{data}
\keyword{dataframe.}
\keyword{file}
\keyword{function}
\keyword{healthcare}
\keyword{it}
\keyword{reads}
\keyword{returns}
\keyword{trial}
