% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical_imaging_module.R
\name{load_medical_imaging_data}
\alias{load_medical_imaging_data}
\title{Medical Imaging Module}
\usage{
load_medical_imaging_data(path)
}
\arguments{
\item{path}{A character string specifying the file path to the medical
imaging dataset in PNG format.}
}
\value{
A 3D array representing the medical imaging data. Each layer
corresponds to a color channel.
}
\description{
This module covers techniques for medical imaging analysis, including image
segmentation, object detection, and radiological image interpretation.
}
\examples{
medical_data <- load_medical_imaging_data("path/to/your/image.png")
}
\author{
Khawla Aghzawi
}
\keyword{Data}
\keyword{Imaging}
\keyword{Load}
\keyword{Medical}
\keyword{PNG}
\keyword{This}
\keyword{a}
\keyword{analysis}
\keyword{data}
\keyword{dataset}
\keyword{file}
\keyword{format}
\keyword{from}
\keyword{function}
\keyword{healthcare}
\keyword{imaging}
\keyword{in}
\keyword{medical}
\keyword{path.}
\keyword{reads}
\keyword{specified}
\keyword{the}
