% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_models_module.R
\name{load_predictive_models_data}
\alias{load_predictive_models_data}
\title{Predictive Models Module}
\usage{
load_predictive_models_data()
}
\value{
A dataframe containing predictive models data.
}
\description{
This module covers building predictive models for healthcare data, including
patient outcome predictions.
}
\author{
Khawla Aghzawi
}
\keyword{Data}
\keyword{Load}
\keyword{Models}
\keyword{Predictive}
\keyword{This}
\keyword{a}
\keyword{building}
\keyword{data}
\keyword{dataset}
\keyword{for}
\keyword{function}
\keyword{healthcare}
\keyword{loads}
\keyword{modeling}
\keyword{models.}
\keyword{predictive}
\keyword{sample}
