% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_models_module.R
\name{make_predictions}
\alias{make_predictions}
\title{Make Predictions}
\usage{
make_predictions(trained_model, new_data)
}
\arguments{
\item{trained_model}{A trained predictive model.}

\item{new_data}{A dataframe containing new data for predictions.}
}
\value{
Predictions for the new data.
}
\description{
This function makes predictions using a trained predictive model.
}
