% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_trial_module.R
\name{perform_hypothesis_test}
\alias{perform_hypothesis_test}
\title{Perform Hypothesis Test on Clinical Trial Data}
\usage{
perform_hypothesis_test(clinical_trial_data)
}
\arguments{
\item{clinical_trial_data}{A dataframe containing clinical trial data with columns 'Treatment' and 'Outcome'.}
}
\value{
The result of the chi-square test for independence.
}
\description{
This function performs a chi-square test for independence on the given clinical trial data.
}
\examples{
  clinical_data <- load_clinical_trial_data("path/to/clinical_data.csv")
  perform_hypothesis_test(clinical_data)

}
