% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical_imaging_module.R
\name{perform_image_segmentation}
\alias{perform_image_segmentation}
\title{Perform Image Segmentation}
\usage{
perform_image_segmentation(image)
}
\arguments{
\item{image}{An image for segmentation.}
}
\value{
A segmented image.
}
\description{
This function performs image segmentation on medical imaging data.
}
