% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical_imaging_module.R
\name{perform_object_detection}
\alias{perform_object_detection}
\title{Perform Object Detection on Breast Ultrasound Images}
\usage{
perform_object_detection(
  dataset_path,
  subset_size = NULL,
  dimensions = c(256, 256),
  test_split = 0.2,
  validation_split = 0.1,
  epochs = 20,
  batch_size = 8,
  threshold = 0.8
)
}
\arguments{
\item{dataset_path}{Path to the directory containing breast ultrasound images.}

\item{dimensions}{Dimensions to which the images will be resized (default is c(256, 256)).}

\item{test_split}{Fraction of images for testing (default is 0.2).}

\item{validation_split}{Fraction of images for validation (default is 0.1).}

\item{epochs}{Number of training epochs for the model (default is 20).}

\item{batch_size}{Batch size used during model training (default is 8).}

\item{threshold}{Probability threshold for image segmentation (default is 0.8).}
}
\value{
A list containing the results of object detection on test images.
}
\description{
This function facilitates object detection on breast ultrasound images for the task
of identifying tumors. It leverages the \code{imager} package for image processing
and \code{keras} for deep learning model creation.
}
\examples{
dataset_path <- "C:/Users/User/Downloads/Dataset_BUSI_with_GT"
detection_results <- perform_object_detection(dataset_path, 20)

}
\seealso{
\code{\link{imager}} package for image processing.
\code{\link{keras}} package for deep learning model creation.
}
