% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical_imaging_module.R
\name{preprocess_medical_imaging_data}
\alias{preprocess_medical_imaging_data}
\title{Preprocess Medical Imaging Data}
\usage{
preprocess_medical_imaging_data(medical_imaging_data)
}
\arguments{
\item{medical_imaging_data}{A 2D or 3D array representing medical imaging data.}
}
\value{
Preprocessed medical imaging data.
}
\description{
This function performs preprocessing on medical imaging data.
}
