% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_models_module.R
\name{preprocess_predictive_models_data}
\alias{preprocess_predictive_models_data}
\title{Preprocess Predictive Models Data}
\usage{
preprocess_predictive_models_data(predictive_models_data)
}
\arguments{
\item{predictive_models_data}{A dataframe containing predictive models data.}
}
\value{
A preprocessed dataframe.
}
\description{
This function preprocesses predictive models data.
}
