% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_trial_module.R
\name{visualize_clinical_trial_data}
\alias{visualize_clinical_trial_data}
\title{Visualize Clinical Trial Data}
\usage{
visualize_clinical_trial_data(clinical_trial_data)
}
\arguments{
\item{clinical_trial_data}{A dataframe containing preprocessed clinical trial data.}
}
\description{
This function creates visualizations for clinical trial data, such as histograms, bar charts, etc.
}
