% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical_imaging_module.R
\name{visualize_medical_imaging_data}
\alias{visualize_medical_imaging_data}
\title{Visualize Medical Imaging Data}
\usage{
visualize_medical_imaging_data(medical_imaging_data)
}
\arguments{
\item{medical_imaging_data}{A list containing preprocessed medical imaging data.}
}
\description{
This function creates visualizations for medical imaging data.
}
