% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_module.R
\name{analyze_ehr_data}
\alias{analyze_ehr_data}
\title{EHR Module}
\usage{
analyze_ehr_data(ehr_data)
}
\arguments{
\item{ehr_data}{to dataset}
}
\value{
a summary staatistics and plots
}
\description{
This module focuses on analyzing and visualizing EHR data.
}
\examples{
analyze_ehr_data("PATH/TO/data/ehr_sample_data.csv")
}
\author{
Khawla Aghzawi
}
\keyword{Analyze}
\keyword{EHR}
\keyword{analysis}
\keyword{data}
\keyword{ehr}
\keyword{healthcare}
