% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medical_imaging_module.R
\name{perform_object_detection}
\alias{perform_object_detection}
\title{Perform Object Detection}
\usage{
perform_object_detection(image)
}
\arguments{
\item{image}{An image for object detection.}
}
\value{
A list containing detected objects.
}
\description{
This function performs object detection on medical imaging data.
}
